+++ packages/use/src/whenever/index.test.ts
@@ -1,10 +1,123 @@
 import { nextTick, ref } from 'vue-demi';
-import { whenever as vueuseWhenever } from '@vueuse/core';
 import { whenever, wheneverDeep, wheneverImmediate, wheneverImmediateDeep } from '@mixte/use';
 
 describe('whenever', () => {
-  it('导出的 whenever 方法为 @vueuse/core 中的 whenever 方法', () => {
-    expect(whenever).toBe(vueuseWhenever);
+  it('监听值是 `truthy` 时执行回调函数', async () => {
+    const a = ref<any>(1);
+    const watchFns = {
+      fn1: () => {},
+      fn2: () => {},
+      fn3: () => {},
+      fn4: () => {},
+      fn5: () => {},
+      fn6: () => {},
+      fn7: () => {},
+      fn8: () => {},
+    };
+
+    const fn1 = vi.spyOn(watchFns, 'fn1');
+    const fn2 = vi.spyOn(watchFns, 'fn2');
+    const fn3 = vi.spyOn(watchFns, 'fn3');
+    const fn4 = vi.spyOn(watchFns, 'fn4');
+    const fn5 = vi.spyOn(watchFns, 'fn5');
+    const fn6 = vi.spyOn(watchFns, 'fn6');
+    const fn7 = vi.spyOn(watchFns, 'fn7');
+    const fn8 = vi.spyOn(watchFns, 'fn8');
+
+    whenever(a, watchFns.fn1);
+    whenever(a, watchFns.fn2, { immediate: true });
+    whenever(a, watchFns.fn3, { immediate: true, deep: true });
+    whenever(a, watchFns.fn4, { immediate: true, once: true });
+    whenever(a, watchFns.fn5, { immediate: true, deep: true, once: true });
+    whenever(a, watchFns.fn6, { deep: true });
+    whenever(a, watchFns.fn7, { deep: true, once: true });
+    whenever(a, watchFns.fn8, { once: true });
+
+    // 1: truthy
+    expect(fn1).not.toHaveBeenCalled();
+    expect(fn2).toHaveBeenCalledTimes(1); // immediate
+    expect(fn3).toHaveBeenCalledTimes(1); // immediate + deep
+    expect(fn4).toHaveBeenCalledTimes(1); // immediate + once
+    expect(fn5).toHaveBeenCalledTimes(1); // immediate + deep + once
+    expect(fn6).not.toHaveBeenCalled(); // deep
+    expect(fn7).not.toHaveBeenCalled(); // deep + once
+    expect(fn8).not.toHaveBeenCalled(); // once
+
+    await nextTick();
+
+    expect(fn1).not.toHaveBeenCalled();
+    expect(fn2).toHaveBeenCalledTimes(1); // immediate
+    expect(fn3).toHaveBeenCalledTimes(1); // immediate + deep
+    expect(fn4).toHaveBeenCalledTimes(1); // immediate + once
+    expect(fn5).toHaveBeenCalledTimes(1); // immediate + deep + once
+    expect(fn6).not.toHaveBeenCalled(); // deep
+    expect(fn7).not.toHaveBeenCalled(); // deep + once
+    expect(fn8).not.toHaveBeenCalled(); // once
+
+    // 2: truthy
+    a.value = 2;
+    await nextTick();
+
+    expect(fn1).toHaveBeenCalledTimes(1);
+    expect(fn2).toHaveBeenCalledTimes(2); // immediate
+    expect(fn3).toHaveBeenCalledTimes(2); // immediate + deep
+    expect(fn4).toHaveBeenCalledTimes(1); // immediate + once
+    expect(fn5).toHaveBeenCalledTimes(1); // immediate + deep + once
+    expect(fn6).toHaveBeenCalledTimes(1); // deep
+    expect(fn7).toHaveBeenCalledTimes(1); // deep + once
+    expect(fn8).toHaveBeenCalledTimes(1); // once
+
+    // 3: truthy
+    a.value = 3;
+    await nextTick();
+
+    expect(fn1).toHaveBeenCalledTimes(2);
+    expect(fn2).toHaveBeenCalledTimes(3); // immediate
+    expect(fn3).toHaveBeenCalledTimes(3); // immediate + deep
+    expect(fn4).toHaveBeenCalledTimes(1); // immediate + once
+    expect(fn5).toHaveBeenCalledTimes(1); // immediate + deep + once
+    expect(fn6).toHaveBeenCalledTimes(2); // deep
+    expect(fn7).toHaveBeenCalledTimes(1); // deep + once
+    expect(fn8).toHaveBeenCalledTimes(1); // once
+
+    // 0: falsy
+    a.value = 0;
+    await nextTick();
+
+    expect(fn1).toHaveBeenCalledTimes(2);
+    expect(fn2).toHaveBeenCalledTimes(3); // immediate
+    expect(fn3).toHaveBeenCalledTimes(3); // immediate + deep
+    expect(fn4).toHaveBeenCalledTimes(1); // immediate + once
+    expect(fn5).toHaveBeenCalledTimes(1); // immediate + deep + once
+    expect(fn6).toHaveBeenCalledTimes(2); // deep
+    expect(fn7).toHaveBeenCalledTimes(1); // deep + once
+    expect(fn8).toHaveBeenCalledTimes(1); // once
+
+    // { a: 1 }: truthy
+    a.value = { a: 1 };
+    await nextTick();
+
+    expect(fn1).toHaveBeenCalledTimes(3);
+    expect(fn2).toHaveBeenCalledTimes(4); // immediate
+    expect(fn3).toHaveBeenCalledTimes(4); // immediate + deep
+    expect(fn4).toHaveBeenCalledTimes(1); // immediate + once
+    expect(fn5).toHaveBeenCalledTimes(1); // immediate + deep + once
+    expect(fn6).toHaveBeenCalledTimes(3); // deep
+    expect(fn7).toHaveBeenCalledTimes(1); // deep + once
+    expect(fn8).toHaveBeenCalledTimes(1); // once
+
+    // { a: 2 }: truthy
+    a.value.a = 2;
+    await nextTick();
+
+    expect(fn1).toHaveBeenCalledTimes(3);
+    expect(fn2).toHaveBeenCalledTimes(4); // immediate
+    expect(fn3).toHaveBeenCalledTimes(5); // immediate + deep
+    expect(fn4).toHaveBeenCalledTimes(1); // immediate + once
+    expect(fn5).toHaveBeenCalledTimes(1); // immediate + deep + once
+    expect(fn6).toHaveBeenCalledTimes(4); // deep
+    expect(fn7).toHaveBeenCalledTimes(1); // deep + once
+    expect(fn8).toHaveBeenCalledTimes(1); // once
   });
 });