+++ packages/use/src/whenever/index.ts
@@ -1,5 +1,25 @@
+import type { WheneverOptions } from '@vueuse/core';
 import type { WatchCallback, WatchOptions, WatchSource } from 'vue-demi';
-import { whenever } from '@vueuse/core';
+import { nextTick, watch } from 'vue-demi';
+
+/**
+ * 监听值是 `truthy` 时执行回调函数
+ *
+ * @see https://mixte.moomfe.com/mixte/use/whenever#whenever
+ */
+export function whenever<T>(source: WatchSource<T | false | null | undefined>, cb: WatchCallback<T>, options?: WheneverOptions) {
+  const stop = watch(
+    source,
+    (v, ov, onCleanup) => {
+      if (v) {
+        options?.once && nextTick(() => stop());
+        cb(v, ov, onCleanup);
+      }
+    },
+    { ...options, once: false },
+  );
+  return stop;
+}
 
 /**
  * immediate 默认为 true 的 whenever 方法
@@ -27,5 +47,3 @@ export function wheneverDeep<T>(source: WatchSource<T | false | null | undefined
 export function wheneverImmediateDeep<T>(source: WatchSource<T | false | null | undefined>, cb: WatchCallback<T>, options?: Omit<WatchOptions, 'immediate' | 'deep'>) {
   return whenever(source, cb, { ...options, immediate: true, deep: true });
 }
-
-export { whenever } from '@vueuse/core';